+++ src/bika/lims/browser/auditlog.py
@@ -27,13 +27,13 @@ from bika.lims.api.snapshot import get_snapshot_by_version
 from bika.lims.api.snapshot import get_snapshot_metadata
 from bika.lims.api.snapshot import get_snapshot_version
 from bika.lims.api.snapshot import get_snapshots
-from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.interfaces import IAuditable
 from bika.lims.utils import t
 from plone.memoize import view
 from Products.CMFPlone.i18nl10n import ulocalized_time
 from Products.Five.browser import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from senaite.app.listing import ListingView
 
 
 class HasAuditLog(BrowserView):
@@ -43,7 +43,7 @@ class HasAuditLog(BrowserView):
         return IAuditable.providedBy(self.context)
 
 
-class AuditLogView(BikaListingView):
+class AuditLogView(ListingView):
     """Audit View
     """
     diff_template = ViewPageTemplateFile("templates/auditlog_diff.pt")
@@ -55,6 +55,7 @@ class AuditLogView(BikaListingView):
         # the default settings
         self.catalog = "uid_catalog"
         self.contentFilter = {"UID": api.get_uid(context)}
+        self.context_actions = {}
 
         # TODO: Fix in senaite.app.listing.get_api_url
         #