+++ src/bika/lims/browser/widgets/artemplateanalyseswidget.py
@@ -254,12 +254,9 @@ class ARTemplateAnalysesView(BikaListingView):
         if obj.getAccredited():
             after_icons += get_image(
                 "accredited.png", title=t(_("Accredited")))
-        if obj.getAttachmentOption() == "r":
+        if obj.getAttachmentRequired():
             after_icons += get_image(
                 "attach_reqd.png", title=t(_("Attachment required")))
-        if obj.getAttachmentOption() == "n":
-            after_icons += get_image(
-                "attach_no.png", title=t(_('Attachment not permitted')))
         if after_icons:
             item["after"]["Title"] = after_icons