+++ src/bika/lims/vocabularies/__init__.py
@@ -32,7 +32,6 @@ from plone.resource.utils import iterDirectoriesOfType
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
-from zope.component import getAdapters
 from zope.site.hooks import getSite
 
 import os