+++ src/bika/lims/workflow/worksheet/events.py
@@ -21,20 +21,6 @@
 from bika.lims import workflow as wf
 
 
-def after_submit(obj):
-    """
-    Method triggered after a 'submit' transition for the Worksheet passed in is
-    performed.
-    This function is called automatically by
-    bika.lims.workfow.AfterTransitionEventHandler
-    """
-    # Submitting a Worksheet must never transition the analyses.
-    # In fact, a worksheet can only be transitioned to "to_be_verified" if
-    # all the analyses that contain have been submitted manually after
-    # the results input
-    wf.doActionFor(obj, 'attach')
-
-
 def after_retract(worksheet):
     """Retracts all analyses the worksheet contains
     """