+++ src/senaite/core/exportimport/instruments/panalytical/omnia/axios_xrf.py
@@ -55,11 +55,11 @@ def Import(context, request):
 
     if parser:
         # Load the importer
-        status = ['sample_received', 'attachment_due', 'to_be_verified']
+        status = ['sample_received', 'to_be_verified']
         if artoapply == 'received':
             status = ['sample_received']
         elif artoapply == 'received_tobeverified':
-            status = ['sample_received', 'attachment_due', 'to_be_verified']
+            status = ['sample_received', 'to_be_verified']
 
         over = [False, False]
         if override == 'nooverride':